requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
python-i18n==0.3.9
PyYAML==5.4.1
requests==2.25.1
termcolor==1.1.0
urllib3==1.26.3
